scraper-ddb/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scraper-ddb"
version = "0.1.0"
description = "Scraper for Deutsche Digitale Bibliothek / Archivportal-D"
requires-python = ">=3.13"
dependencies = [
    "worker-common",
    "httpx>=0.27",
    "Pillow>=10.0",
    "reportlab>=4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-httpx>=0.30",
]

[project.scripts]
scraper-ddb = "scraper_ddb.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/scraper_ddb"]
